requirements.txt
PyPDF2
streamlit
google-generativeai
langchain
PyPDF2
chromadb
faiss-cpu
langchain_google_genai
langchain-community
